* Backup now uses GMT time instead of local machine * Several bugfixes during activation
git-svn-id: https://plugins.svn.wordpress.org/6scan-protection/trunk@575931 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ modules/backup/backup_func.php
@@ -9,7 +9,7 @@ function sixscan_backup_func_controller( $backup_type , &$backup_err_descr ){
 	/* If there was a false in prerequisites */
 	if ( array_search( FALSE , array_values( $run_backup_prerequisites ) , TRUE ) !== FALSE ){
 		# return prerequisites only for database
-		if ( $backup_type == SIXSCAN_BACKUP_DATABASE_REQUEST ) $backup_err_descr[ 'prerequisites' ] = $run_backup_prerequisites;	
+		$backup_err_descr[ 'prerequisites' ] = $run_backup_prerequisites;	
 		$backup_err_descr[ 'success' ] = FALSE;
 		return FALSE;
 	}
@@ -195,10 +195,16 @@ function sixscan_backup_func_db( &$backed_up_filename ){
 	/*	If a previous file was not deleted from some reason, delete it now and save the current filename */
 	sixscan_backup_func_delete_previous( SIXSCAN_BACKUP_LAST_DB_NAME , $temp_sql_file_zipped );		
 
-	sixscan_backup_sql( DB_HOST , DB_USER , DB_PASSWORD , DB_NAME , $temp_sql_file_name );
+	$sql_backup_result = sixscan_backup_sql( DB_HOST , DB_USER , DB_PASSWORD , DB_NAME , $temp_sql_file_name );
+	if ( $sql_backup_result != TRUE ){
+		$backed_up_filename[ 'internal_message' ] = "Could not create SQL file. Error returned: $sql_backup_result";
+		$backed_up_filename[ 'user_result_message' ] = "Failed creating SQL backup (Reported error: $sql_backup_result )";
+		$backed_up_filename[ 'success' ] = FALSE;		
+		return FALSE;				
+	}
 	
 	$zip_result = sixscan_backup_func_compress( $temp_sql_file_name, $temp_sql_file_zipped );
-	if ($zip_result !== TRUE)
+	if ( $zip_result !== TRUE )
 	{
 		$backed_up_filename[ 'internal_message' ] = "Could not create zip file. Error returned: $zip_result";
 		$backed_up_filename[ 'user_result_message' ] = "Failed creating ZIP file (write permissions at " . get_temp_dir() . " ?)";
@@ -230,12 +236,18 @@ function sixscan_backup_sql( $host , $user , $pass , $name , $sql_output_file )
   	define( 'DUMP_RECORD_COUNT' , 1000 ); 
 
   	/* Access the SQL database */
-	$link = mysql_connect( $host , $user , $pass );
-	mysql_select_db( $name , $link );
+	$sql_link = mysql_connect( $host , $user , $pass );
+	if ( $sql_link === FALSE )
+		return "Failed connecting to MySQL database on $host";
+
+	mysql_select_db( $name , $sql_link );
   
  	/*	Getting table list */
 	$tables = array();
-	$result = mysql_query( 'SHOW TABLES' );
+	$result = mysql_query( 'SHOW TABLES' , $sql_link );
+	if ( $result == FALSE )
+		return "Failed enumerating tables";
+
 	while( $row = mysql_fetch_row( $result ) )
 	{
 		$tables[] = $row[0];
@@ -244,12 +256,15 @@ function sixscan_backup_sql( $host , $user , $pass , $name , $sql_output_file )
   	$dumped_record_count = 0;
 	$sql_dump_data = '';	
 	foreach( $tables as $table ){
-		$result = mysql_query( 'SELECT * FROM ' . $table );
+		$result = mysql_query( 'SELECT * FROM ' . $table , $sql_link );		
+		if ( $result == FALSE )
+			return "Failed reading table $table";
+
 		$num_fields = mysql_num_fields( $result );
 
 		/* Create tables syntax: */
 		$sql_dump_data.= 'DROP TABLE IF EXISTS `' . $table . '`;';
-		$row2 = mysql_fetch_row( mysql_query( 'SHOW CREATE TABLE `' . $table . '`' ) );
+		$row2 = mysql_fetch_row( mysql_query( 'SHOW CREATE TABLE `' . $table . '`' , $sql_link ) );
 		$sql_dump_data .= "\n\n" . $row2[ 1 ] . ";\n\n";
 
 		/* Dump each table's data */
@@ -281,7 +296,9 @@ function sixscan_backup_sql( $host , $user , $pass , $name , $sql_output_file )
 				$dumped_record_count++;
 				if ( $dumped_record_count == DUMP_RECORD_COUNT ){
 					// append data to avoid large memory demands
-					file_put_contents( $sql_output_file , $sql_dump_data , FILE_APPEND );
+					if ( file_put_contents( $sql_output_file , $sql_dump_data , FILE_APPEND ) === FALSE )
+						return "Failed writing data to SQL backup file $sql_output_file";
+
 					$sql_dump_data = '';
 					$dumped_record_count = 0;
 				}
@@ -290,11 +307,15 @@ function sixscan_backup_sql( $host , $user , $pass , $name , $sql_output_file )
 
 		$sql_dump_data.="\n\n\n";
 		// append data to avoid large memory demands
-		file_put_contents( $sql_output_file , $sql_dump_data , FILE_APPEND );
+		if ( file_put_contents( $sql_output_file , $sql_dump_data , FILE_APPEND ) === FALSE ){
+			return "Failed writing data to SQL backup file $sql_output_file";
+		}
+
 		$sql_dump_data = '';
 	}
 	
-	mysql_close( $link );
+	mysql_close( $sql_link );
+	return TRUE;
 }
 
 ?>
